+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add a `vim.g.ferris.server.auto_attach` option, which
+  can be a `boolean` or a `fun():boolean` that determines
+  whether or not to auto-attach the LSP client when opening
+  a Rust file.
+
 ## [2.0.0] - 2023-11-21
 
 ### Breaking changes